+++ CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - On `ChessGame.jumpToMainlinePly`, fixed not getting fen for `notifyPositionJumped` inside of `writeLock`, but outside of `writeLock`.
 - On `ChessGame.setCurrentMoveClockMilliSeconds`, added zero-padding
 - On `ConvertStringMoveUtils.lanToMoveData`, fixed verifying square data too late
+- On `ChessGame.loadPGN`, fixed checking `gameResult`, `gameOverReason`, `fen` outside of `writeLock`.
 
 ### Performance
 - On `MoveGenerator`, replaced `if` cases on piece attack finding to `switch-case`.